+++ modules/configuration.nix
@@ -24,6 +24,13 @@
   # networking.proxy.default = "http://user:password@proxy:port/";
   # networking.proxy.noProxy = "127.0.0.1,localhost,internal.domain";
 
+  # Home Manager
+  home-manager = {
+    extraSpecialArgs.inputs = inputs;
+    useGlobalPkgs = true;
+    useUserPackages = true;
+  };
+
   # Enable networking
   networking.networkmanager.enable = true;